--- app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/EmailConfirmationViewModel.kt
@@ -1,4 +1,0 @@
-package com.margaritalashina.mytrainingprojectmobiledev
-
-class EmailConfirmationViewModel : BaseViewModel() {
-}
