--- app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/MainFragmentViewModel.kt
@@ -1,4 +1,0 @@
-package com.margaritalashina.mytrainingprojectmobiledev
-
-class MainFragmentViewModel : BaseViewModel() {
-}
