--- app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/MainViewModel.kt
@@ -1,3 +1,0 @@
-package com.margaritalashina.mytrainingprojectmobiledev
-
-class MainViewModel : BaseViewModel()
