--- app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/SignInViewModel.kt
@@ -1,4 +1,0 @@
-package com.margaritalashina.mytrainingprojectmobiledev
-
-class SignInViewModel : BaseViewModel() {
-}
