--- app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/SignUpViewModel.kt
@@ -1,4 +1,0 @@
-package com.margaritalashina.mytrainingprojectmobiledev
-
-class SignUpViewModel : BaseViewModel() {
-}
