+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/data/network/Api.kt
@@ -1,5 +1,6 @@
 package com.margaritalashina.mytrainingprojectmobiledev
 
+import com.margaritalashina.mytrainingprojectmobiledev.entity.User
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import retrofit2.http.GET
