+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/MainViewModel.kt
@@ -1,0 +1,5 @@
+package com.margaritalashina.mytrainingprojectmobiledev.ui
+
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseViewModel
+
+class MainViewModel : BaseViewModel()
