+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/base/BaseFragment.kt
@@ -1,8 +1,11 @@
-package com.margaritalashina.mytrainingprojectmobiledev
+package com.margaritalashina.mytrainingprojectmobiledev.ui.base
 
 import androidx.annotation.LayoutRes
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.margaritalashina.mytrainingprojectmobiledev.BuildConfig
+import com.margaritalashina.mytrainingprojectmobiledev.logBackstack
+import com.margaritalashina.mytrainingprojectmobiledev.logFragmentHierarchy
 import timber.log.Timber
 
 open class BaseFragment : Fragment {
