+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/confirmation/EmailConfirmationViewModel.kt
@@ -1,0 +1,6 @@
+package com.margaritalashina.mytrainingprojectmobiledev.ui.confirmation
+
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseViewModel
+
+class EmailConfirmationViewModel : BaseViewModel() {
+}
