+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/main/MainFragment.kt
@@ -1,7 +1,9 @@
-package com.margaritalashina.mytrainingprojectmobiledev
+package com.margaritalashina.mytrainingprojectmobiledev.ui.main
 
 import androidx.fragment.app.viewModels
 import by.kirich1409.viewbindingdelegate.viewBinding
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseFragment
+import com.margaritalashina.mytrainingprojectmobiledev.R
 import com.margaritalashina.mytrainingprojectmobiledev.databinding.FragmentMainBinding
 
 class MainFragment : BaseFragment(R.layout.fragment_main) {
