+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/main/MainFragmentViewModel.kt
@@ -1,0 +1,6 @@
+package com.margaritalashina.mytrainingprojectmobiledev.ui.main
+
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseViewModel
+
+class MainFragmentViewModel : BaseViewModel() {
+}
