+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/onboarding/OnboardingAdapterDelegates.kt
