+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/signin/SignInFragment.kt
@@ -1,7 +1,9 @@
-package com.margaritalashina.mytrainingprojectmobiledev
+package com.margaritalashina.mytrainingprojectmobiledev.ui.signin
 
 import androidx.fragment.app.viewModels
 import by.kirich1409.viewbindingdelegate.viewBinding
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseFragment
+import com.margaritalashina.mytrainingprojectmobiledev.R
 import com.margaritalashina.mytrainingprojectmobiledev.databinding.FragmentSignInBinding
 
 class SignInFragment : BaseFragment(R.layout.fragment_sign_in) {
