+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/signin/SignInViewModel.kt
@@ -1,0 +1,6 @@
+package com.margaritalashina.mytrainingprojectmobiledev.ui.signin
+
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseViewModel
+
+class SignInViewModel : BaseViewModel() {
+}
