+++ app/src/main/java/com/margaritalashina/mytrainingprojectmobiledev/ui/signup/SignUpViewModel.kt
@@ -1,0 +1,6 @@
+package com.margaritalashina.mytrainingprojectmobiledev.ui.signup
+
+import com.margaritalashina.mytrainingprojectmobiledev.ui.base.BaseViewModel
+
+class SignUpViewModel : BaseViewModel() {
+}
