+++ app/src/main/res/layout/activity_main.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/fragmentContainerView"
-    android:name="com.margaritalashina.mytrainingprojectmobiledev.OnboardingFragment"
+    android:name="com.margaritalashina.mytrainingprojectmobiledev.ui.onboarding.OnboardingFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
